Renamed files in the database folder to not use DB or database. Fixed all import statement as well
